+++ packages/idx/src/member/details.component.ts
@@ -1,4 +1,4 @@
-// IdxMemberDetails Component
+// IdxMemberDetails Component  @stratusjs/idx/member/details.component
 // --------------
 
 // Runtime
@@ -15,7 +15,7 @@ import moment from 'moment'
 
 // Services
 // import {Collection} from 'stratus.services.collection' // TODO not sure how to resolve type Promise<Collection>
-import 'stratus.services.idx' // import '@stratusjs/idx/services/idx' The reference for later
+import '@stratusjs/idx/idx'
 
 // FIXME move filters to @stratusjs
 // Custom Filters
