+++ packages/idx/src/property/list.component.ts
@@ -1,4 +1,4 @@
-// IdxPropertyList Component
+// IdxPropertyList Component @stratusjs/idx/property/list.component
 // --------------
 
 // Runtime
@@ -15,10 +15,11 @@ import moment from 'moment'
 
 // Services
 // import {Collection} from 'stratus.services.collection' // TODO not sure how to resolve type Promise<Collection>
-import 'stratus.services.idx' // import '@stratusjs/idx/services/idx' The reference for later
+import '@stratusjs/idx/idx'
 
 // Component Preload
-import 'stratus.components.propertyDetails'
+// import 'stratus.components.propertyDetails'
+import '@stratusjs/idx/property/details.component'
 
 // Stratus Dependencies
 import {isJSON} from '@stratusjs/core/misc'
